Backend/requirements.txt
Django==5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
drf-yasg
python-dotenv
psycopg2-binary
